@@ -8,8 +8,8 @@ import de.johannaherrmann.javauno.exceptions.InvalidTokenException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
-import javax.servlet.RequestDispatcher;
-import javax.servlet.http.HttpServletRequest;
+import jakarta.servlet.RequestDispatcher;
+import jakarta.servlet.http.HttpServletRequest;
 
 class ErrorResponseUtil {
     static ResponseEntity<GeneralResponse> getExceptionResponseEntity(Exception exception){
@@ -29,29 +29,20 @@ class ErrorResponseUtil {
 
     static HttpStatus getErrorStatus(HttpServletRequest request){
         Object status = request.getAttribute(RequestDispatcher.ERROR_STATUS_CODE);
-        Integer statusCode = 500;
+        int statusCode = 500;
         if(status != null) {
-            statusCode = Integer.valueOf(status.toString());
+            statusCode = (Integer) status;
         }
         return HttpStatus.resolve(statusCode);
     }
 
     static ErrorResponse getErrorResponse(HttpStatus httpStatus){
-        String error;
-        switch (httpStatus){
-            case NOT_FOUND:
-                error = "Bad URL. Wrong path or missing url arguments";
-                break;
-            case METHOD_NOT_ALLOWED:
-                error = "Wrong request method";
-                break;
-            case UNSUPPORTED_MEDIA_TYPE:
-                error = "Unsupported media type";
-                break;
-            default:
-                error = "Unknown error";
-                break;
-        }
+        String error = switch (httpStatus) {
+            case NOT_FOUND -> "Bad URL. Wrong path or missing url arguments";
+            case METHOD_NOT_ALLOWED -> "Wrong request method";
+            case UNSUPPORTED_MEDIA_TYPE -> "Unsupported media type";
+            default -> "Unknown error";
+        };
         return new ErrorResponse(error);
     }
 }
